# web/src/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "GameEvent.Web | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/test/clock": {
      "get": {
        "tags": [
          "Test"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestClockView"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Test"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/TestClockRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestClockView"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/test/random": {
      "post": {
        "tags": [
          "Test"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestRandomRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/test/seasons/{seasonId}/scenarios/{name}": {
      "post": {
        "tags": [
          "Test"
        ],
        "parameters": [
          {
            "name": "seasonId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestScenarioView"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "TestClockRequest": {
        "type": "object",
        "properties": {
          "advanceMinutes": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "moveTo": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "reset": {
            "type": "boolean",
            "default": false
          }
        },
        "description": "Move the clock: forward by minutes, to a moment, or back to the real time."
      },
      "TestClockView": {
        "required": [
          "now",
          "adjustable",
          "shiftMinutes"
        ],
        "type": "object",
        "properties": {
          "now": {
            "type": "string",
            "format": "date-time"
          },
          "adjustable": {
            "type": "boolean"
          },
          "shiftMinutes": {
            "type": "number",
            "format": "double"
          }
        },
        "description": "The site's clock as the test endpoints see it: now, whether it can be moved and how far it is from the real time\n(`shiftMinutes`, negative — in the past; H9, D-221)."
      },
      "TestRandomRequest": {
        "required": [
          "seed"
        ],
        "type": "object",
        "properties": {
          "seed": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          }
        },
        "description": "Seed the randomness (the same rolls and dice every time), or null for unpredictable randomness again."
      },
...
      "TestScenarioRequest": {
        "type": "object",
        "properties": {
          "player": {
            "type": [
              "null",
              "string"
            ]
          },
          "playerId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        },
        "description": "A scenario for a player of the season: by the player's id in the season (`playerId`, the page picks from the\nseason's players) or by login (`player`); the first player by name when neither."
      },
      "TestScenarioView": {
        "required": [
          "scenario",
          "commands"
        ],
        "type": "object",
        "properties": {
          "scenario": {
            "type": "string"
          },
          "commands": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "description": "What a scenario did: the commands it sent, in order."
      },
...
    }
  },
...
}